+++ api.v7/auth/qbox/qbox.go
@@ -1,0 +1,11 @@
+package qbox
+
+import (
+	"net/http"
+	"time"
+)
+
+func NewClient(mac *Mac, transport http.RoundTripper) *http.Client {
+	t := NewTransport(mac, transport)
+	return &http.Client{Transport: t, Timeout: 10 * time.Minute}
+}
