+++ api.v7/conf/conf.go
@@ -9,29 +9,54 @@ import (
 	"github.com/service-sdk/go-sdk-qn/x/rpc.v7"
 )
 
-var version = "1.4.0"
-
-var ACCESS_KEY string
-var SECRET_KEY string
-
-// ----------------------------------------------------------
-
 const (
-	ctypeAppName = ctype.ALPHA | ctype.DIGIT | ctype.UNDERLINE | ctype.SPACE_BAR | ctype.SUB | ctype.DOT
+	version = "1.4.0"
 )
 
-// userApp should be [A-Za-z0-9_\ \-\.]*
-func SetAppName(userApp string) error {
-	if userApp != "" && !ctype.IsType(ctypeAppName, userApp) {
+// 全局的ak，sk
+var (
+	AccessKey string
+	SecretKey string
+)
+
+// checkAppName userAppName should be [A-Za-z0-9_\ \-\.]*
+func checkAppName(appName string) error {
+	const ctypeAppName = ctype.ALPHA | ctype.DIGIT | ctype.UNDERLINE | ctype.SPACE_BAR | ctype.SUB | ctype.DOT
+	if appName != "" && !ctype.IsType(ctypeAppName, appName) {
 		return syscall.EINVAL
 	}
-	rpc.UserAgent = fmt.Sprintf(
-		"QiniuGo/%s (%s; %s; %s) %s", version, runtime.GOOS, runtime.GOARCH, userApp, runtime.Version())
+	return nil
+}
+
+func uaString(userAppName string) string {
+	return fmt.Sprintf(
+		"QiniuGo/%s (%s; %s; %s) %s",
+		version,
+		runtime.GOOS,
+		runtime.GOARCH,
+		userAppName,
+		runtime.Version(),
+	)
+}
+
+// SetAppName userAppName should be [A-Za-z0-9_\ \-\.]*
+func SetAppName(userAppName string, client *rpc.Client) error {
+	if err := checkAppName(userAppName); err != nil {
+		return err
+	}
+	ua := uaString(userAppName)
+	client.UserAgent = &ua
+	return nil
+}
+
+func SetDefaultAppName(userAppName string) error {
+	if err := checkAppName(userAppName); err != nil {
+		return err
+	}
+	rpc.DefaultUserAgent = uaString(userAppName)
 	return nil
 }
 
 func init() {
-	SetAppName("")
+	_ = SetDefaultAppName("")
 }
-
-// ----------------------------------------------------------
