+++ api.v7/kodo/token.go
@@ -7,21 +7,16 @@ import (
 	"strings"
 	"time"
 
-	"github.com/service-sdk/go-sdk-qn/api.v7/auth/qbox"
 	"github.com/service-sdk/go-sdk-qn/x/url.v7"
 )
 
-// ----------------------------------------------------------
-
-// 根据空间(Bucket)的域名，以及文件的 key，获得 baseUrl。
+// MakeBaseUrl 根据空间(Bucket)的域名，以及文件的 key，获得 baseUrl。
 // 如果空间是 public 的，那么通过 baseUrl 可以直接下载文件内容。
 // 如果空间是 private 的，那么需要对 baseUrl 进行私有签名得到一个临时有效的 privateUrl 进行下载。
 func MakeBaseUrl(domain, key string) (baseUrl string) {
 
 	return "http://" + domain + "/" + url.Escape(key)
 }
-
-// ----------------------------------------------------------
 
 type GetPolicy struct {
 	Expires uint32
@@ -44,41 +39,9 @@ func (p *Client) MakePrivateUrl(baseUrl string, policy *GetPolicy) (privateUrl s
 	}
 	baseUrl += strconv.FormatInt(deadline, 10)
 
-	token := qbox.Sign(p.mac, []byte(baseUrl))
+	token := p.mac.Sign([]byte(baseUrl))
 	return baseUrl + "&token=" + token
 }
-
-// --------------------------------------------------------------------------------
-
-type PutPolicy struct {
-	Scope               string   `json:"scope"`
-	Expires             uint32   `json:"deadline"`             // 截止时间（以秒为单位）
-	InsertOnly          uint16   `json:"insertOnly,omitempty"` // 若非0, 即使Scope为 Bucket:Key 的形式也是insert only
-	DetectMime          uint8    `json:"detectMime,omitempty"` // 若非0, 则服务端根据内容自动确定 MimeType
-	CallbackFetchKey    uint8    `json:"callbackFetchKey,omitempty"`
-	FsizeLimit          int64    `json:"fsizeLimit,omitempty"`
-	MimeLimit           string   `json:"mimeLimit,omitempty"`
-	SaveKey             string   `json:"saveKey,omitempty"`
-	CallbackUrl         string   `json:"callbackUrl,omitempty"`
-	CallbackHost        string   `json:"callbackHost,omitempty"`
-	CallbackBody        string   `json:"callbackBody,omitempty"`
-	CallbackBodyType    string   `json:"callbackBodyType,omitempty"`
-	ReturnUrl           string   `json:"returnUrl,omitempty"`
-	ReturnBody          string   `json:"returnBody,omitempty"`
-	PersistentOps       string   `json:"persistentOps,omitempty"`
-	PersistentNotifyUrl string   `json:"persistentNotifyUrl,omitempty"`
-	PersistentPipeline  string   `json:"persistentPipeline,omitempty"`
-	AsyncOps            string   `json:"asyncOps,omitempty"`
-	EndUser             string   `json:"endUser,omitempty"`
-	Checksum            string   `json:"checksum,omitempty"` // 格式：<HashName>:<HexHashValue>，目前支持 MD5/SHA1。
-	UpHosts             []string `json:"uphosts,omitempty"`
-	NotifyQueue         string   `json:"notifyQueue,omitempty"`
-	NotifyMessage       string   `json:"notifyMessage,omitempty"`
-	NotifyMessageType   string   `json:"notifyMessageType,omitempty"`
-	DeleteAfterDays     int      `json:"deleteAfterDays,omitempty"`
-	FileType            FileType `json:"fileType,omitempty"`
-}
-
 func (p *Client) MakeUptoken(policy *PutPolicy) string {
 	token, err := p.MakeUptokenWithSafe(policy)
 	if err != nil {
@@ -94,14 +57,6 @@ func (p *Client) MakeUptokenWithSafe(policy *PutPolicy) (token string, err error
 	}
 	rr.Expires += uint32(time.Now().Unix())
 	b, _ := json.Marshal(&rr)
-	token = qbox.SignWithData(p.mac, b)
+	token = p.mac.SignWithData(b)
 	return
 }
-
-func getBucketNameFromPutPolicy(policy *PutPolicy) (bucketName string) {
-	scope := policy.Scope
-	bucketName = strings.Split(scope, ":")[0]
-	return
-}
-
-// ----------------------------------------------------------
