+++ api.v7/kodo/upload.go
@@ -14,10 +14,7 @@ type PutExtra kodocli.PutExtra
 type RputExtra kodocli.RputExtra
 type PutRet kodocli.PutRet
 
-// ----------------------------------------------------------
-
 func (p Bucket) makeUptoken(key string) string {
-
 	policy := &PutPolicy{
 		Scope:   p.Name + ":" + key,
 		Expires: 3600,
@@ -37,15 +34,15 @@ func (p Bucket) makeUptokenWithoutKey() string {
 }
 
 func (p Bucket) makeUploader() kodocli.Uploader {
-
-	c := &http.Client{Transport: p.Conn.Transport, Timeout: 10 * time.Minute}
-	return kodocli.Uploader{Conn: rpc.Client{c}, UpHosts: p.Conn.UpHosts}
+	return kodocli.Uploader{
+		Conn: rpc.Client{
+			Client: &http.Client{Transport: p.Conn.Transport, Timeout: 10 * time.Minute},
+		},
+		UpHosts: p.Conn.UpHosts,
+	}
 }
 
-// ----------------------------------------------------------
-
-// 上传一个文件。
-//
+// Put 上传一个文件。
 // ctx     是请求的上下文。
 // ret     是上传成功后返回的数据。返回的是 PutRet 结构。可选，可以传 nil 表示不感兴趣。
 // key     是要上传的文件访问路径。比如："foo/bar.jpg"。注意我们建议 key 不要以 '/' 开头。另外，key 为空字符串是合法的。
@@ -53,29 +50,37 @@ func (p Bucket) makeUploader() kodocli.Uploader {
 // fsize   是要上传的文件大小。
 // extra   是上传的一些可选项。详细见 PutExtra 结构的描述。
 func (p Bucket) Put(
-	ctx context.Context, ret interface{}, key string, data io.Reader, size int64, extra *PutExtra) error {
-
+	ctx context.Context,
+	ret interface{},
+	key string,
+	data io.Reader,
+	size int64,
+	extra *PutExtra,
+) error {
 	uploader := p.makeUploader()
 	uptoken := p.makeUptoken(key)
 	return uploader.Put(ctx, ret, uptoken, key, data, size, (*kodocli.PutExtra)(extra))
 }
 
-// 上传一个文件。自动以文件的 hash 作为文件的访问路径（key）。
-//
+// PutWithoutKey 上传一个文件。自动以文件的 hash 作为文件的访问路径（key）
 // ctx     是请求的上下文。
 // ret     是上传成功后返回的数据。返回的是 PutRet 结构。可选，可以传 nil 表示不感兴趣。
 // data    是文件内容的访问接口（io.Reader）。
 // fsize   是要上传的文件大小。
 // extra   是上传的一些可选项。详细见 PutExtra 结构的描述。
 func (p Bucket) PutWithoutKey(
-	ctx context.Context, ret interface{}, data io.Reader, size int64, extra *PutExtra) error {
-
+	ctx context.Context,
+	ret interface{},
+	data io.Reader,
+	size int64,
+	extra *PutExtra,
+) error {
 	uploader := p.makeUploader()
 	uptoken := p.makeUptokenWithoutKey()
 	return uploader.PutWithoutKey(ctx, ret, uptoken, data, size, (*kodocli.PutExtra)(extra))
 }
 
-// 上传一个文件。
+// PutFile 上传一个文件。
 // 和 Put 不同的只是一个通过提供文件路径来访问文件内容，一个通过 io.Reader 来访问。
 //
 // ctx       是请求的上下文。
@@ -83,14 +88,17 @@ func (p Bucket) PutWithoutKey(
 // localFile 是要上传的文件的本地路径。
 // extra     是上传的一些可选项。详细见 PutExtra 结构的描述。
 func (p Bucket) PutFile(
-	ctx context.Context, ret interface{}, key, localFile string, extra *PutExtra) (err error) {
-
+	ctx context.Context,
+	ret interface{},
+	key, localFile string,
+	extra *PutExtra,
+) error {
 	uploader := p.makeUploader()
 	uptoken := p.makeUptoken(key)
 	return uploader.PutFile(ctx, ret, uptoken, key, localFile, (*kodocli.PutExtra)(extra))
 }
 
-// 上传一个文件。自动以文件的 hash 作为文件的访问路径（key）。
+// PutFileWithoutKey 上传一个文件。自动以文件的 hash 作为文件的访问路径（key）。
 // 和 PutWithoutKey 不同的只是一个通过提供文件路径来访问文件内容，一个通过 io.Reader 来访问。
 //
 // ctx       是请求的上下文。
@@ -98,16 +106,18 @@ func (p Bucket) PutFile(
 // localFile 是要上传的文件的本地路径。
 // extra     是上传的一些可选项。详细见 PutExtra 结构的描述。
 func (p Bucket) PutFileWithoutKey(
-	ctx context.Context, ret interface{}, localFile string, extra *PutExtra) (err error) {
+	ctx context.Context,
+	ret interface{},
+	localFile string,
+	extra *PutExtra,
+) error {
 
 	uploader := p.makeUploader()
 	uptoken := p.makeUptokenWithoutKey()
 	return uploader.PutFileWithoutKey(ctx, ret, uptoken, localFile, (*kodocli.PutExtra)(extra))
 }
 
-// ----------------------------------------------------------
-
-// 上传一个文件，支持断点续传和分块上传。
+// Rput 上传一个文件，支持断点续传和分块上传。
 //
 // ctx     是请求的上下文。
 // ret     是上传成功后返回的数据。如果 uptoken 中没有设置 CallbackUrl 或 ReturnBody，那么返回的数据结构是 PutRet 结构。
@@ -116,29 +126,37 @@ func (p Bucket) PutFileWithoutKey(
 // fsize   是要上传的文件大小。
 // extra   是上传的一些可选项。详细见 RputExtra 结构的描述。
 func (p Bucket) Rput(
-	ctx context.Context, ret interface{}, key string, data io.ReaderAt, size int64, extra *RputExtra) error {
-
+	ctx context.Context,
+	ret interface{},
+	key string,
+	data io.ReaderAt,
+	size int64,
+	extra *RputExtra,
+) error {
 	uploader := p.makeUploader()
 	uptoken := p.makeUptoken(key)
 	return uploader.Rput(ctx, ret, uptoken, key, data, size, (*kodocli.RputExtra)(extra))
 }
 
-// 上传一个文件，支持断点续传和分块上传。自动以文件的 hash 作为文件的访问路径（key）。
-//
+// RputWithoutKey 上传一个文件，支持断点续传和分块上传。自动以文件的 hash 作为文件的访问路径（key）。
 // ctx     是请求的上下文。
 // ret     是上传成功后返回的数据。如果 uptoken 中没有设置 CallbackUrl 或 ReturnBody，那么返回的数据结构是 PutRet 结构。
 // data    是文件内容的访问接口。考虑到需要支持分块上传和断点续传，要的是 io.ReaderAt 接口，而不是 io.Reader。
 // fsize   是要上传的文件大小。
 // extra   是上传的一些可选项。详细见 RputExtra 结构的描述。
 func (p Bucket) RputWithoutKey(
-	ctx context.Context, ret interface{}, data io.ReaderAt, size int64, extra *RputExtra) error {
-
+	ctx context.Context,
+	ret interface{},
+	data io.ReaderAt,
+	size int64,
+	extra *RputExtra,
+) error {
 	uploader := p.makeUploader()
 	uptoken := p.makeUptokenWithoutKey()
 	return uploader.RputWithoutKey(ctx, ret, uptoken, data, size, (*kodocli.RputExtra)(extra))
 }
 
-// 上传一个文件，支持断点续传和分块上传。
+// RputFile 上传一个文件，支持断点续传和分块上传。
 // 和 Rput 不同的只是一个通过提供文件路径来访问文件内容，一个通过 io.ReaderAt 来访问。
 //
 // ctx       是请求的上下文。
@@ -147,14 +165,17 @@ func (p Bucket) RputWithoutKey(
 // localFile 是要上传的文件的本地路径。
 // extra     是上传的一些可选项。详细见 RputExtra 结构的描述。
 func (p Bucket) RputFile(
-	ctx context.Context, ret interface{}, key, localFile string, extra *RputExtra) (err error) {
-
+	ctx context.Context,
+	ret interface{},
+	key, localFile string,
+	extra *RputExtra,
+) error {
 	uploader := p.makeUploader()
 	uptoken := p.makeUptoken(key)
 	return uploader.RputFile(ctx, ret, uptoken, key, localFile, (*kodocli.RputExtra)(extra))
 }
 
-// 上传一个文件，支持断点续传和分块上传。自动以文件的 hash 作为文件的访问路径（key）。
+// RputFileWithoutKey 上传一个文件，支持断点续传和分块上传。自动以文件的 hash 作为文件的访问路径（key）。
 // 和 RputWithoutKey 不同的只是一个通过提供文件路径来访问文件内容，一个通过 io.ReaderAt 来访问。
 //
 // ctx       是请求的上下文。
@@ -162,8 +183,11 @@ func (p Bucket) RputFile(
 // localFile 是要上传的文件的本地路径。
 // extra     是上传的一些可选项。详细见 RputExtra 结构的描述。
 func (p Bucket) RputFileWithoutKey(
-	ctx context.Context, ret interface{}, localFile string, extra *RputExtra) (err error) {
-
+	ctx context.Context,
+	ret interface{},
+	localFile string,
+	extra *RputExtra,
+) error {
 	uploader := p.makeUploader()
 	uptoken := p.makeUptokenWithoutKey()
 	return uploader.RputFileWithoutKey(ctx, ret, uptoken, localFile, (*kodocli.RputExtra)(extra))
