+++ api.v7/kodocli/main.go
@@ -9,47 +9,6 @@ import (
 	"github.com/service-sdk/go-sdk-qn/x/url.v7"
 )
 
-// ----------------------------------------------------------
-
-type zoneConfig struct {
-	UpHosts []string
-}
-
-var zones = []zoneConfig{
-	// z0:
-	{
-		UpHosts: []string{
-			"http://upload.qiniu.com",
-			"http://up.qiniu.com",
-			"-H up.qiniu.com http://183.136.139.16",
-		},
-	},
-	// z1:
-	{
-		UpHosts: []string{
-			"http://upload-z1.qiniu.com",
-			"http://up-z1.qiniu.com",
-			"-H up-z1.qiniu.com http://106.38.227.27",
-		},
-	},
-	// z2 华南机房:
-	{
-		UpHosts: []string{
-			"http://up-z2.qiniu.com",
-			"http://upload-z2.qiniu.com",
-		},
-	},
-	// na0 北美机房:
-	{
-		UpHosts: []string{
-			"http://up-na0.qiniu.com",
-			"http://upload-na0.qiniu.com",
-		},
-	},
-}
-
-// ----------------------------------------------------------
-
 type UploadConfig struct {
 	UpHosts   []string
 	Transport http.RoundTripper
@@ -63,19 +22,13 @@ type Uploader struct {
 	UpHosts []string
 }
 
-func NewUploader(zone int, cfg *UploadConfig) (p Uploader) {
-
+func NewUploader(cfg *UploadConfig) (p Uploader) {
 	var uc UploadConfig
 	if cfg != nil {
 		uc = *cfg
 	}
 	if uc.Scheme != "https" {
 		uc.Scheme = "http"
-	}
-	if len(uc.UpHosts) == 0 {
-		if zone > 0 && zone < len(zones) {
-			uc.UpHosts = zones[zone].UpHosts
-		}
 	}
 
 	p.UpHosts = uc.UpHosts
@@ -84,24 +37,17 @@ func NewUploader(zone int, cfg *UploadConfig) (p Uploader) {
 }
 
 func NewUploaderWithoutZone(cfg *UploadConfig) (p Uploader) {
-	return NewUploader(-1, cfg)
+	return NewUploader(cfg)
 }
 
-// ----------------------------------------------------------
-
-// 根据空间(Bucket)的域名，以及文件的 key，获得 baseUrl。
+// MakeBaseUrl 根据空间(Bucket)的域名，以及文件的 key，获得 baseUrl。
 // 如果空间是 public 的，那么通过 baseUrl 可以直接下载文件内容。
 // 如果空间是 private 的，那么需要对 baseUrl 进行私有签名得到一个临时有效的 privateUrl 进行下载。
 func MakeBaseUrl(domain, key string) (baseUrl string) {
 	return "http://" + domain + "/" + url.Escape(key)
 }
 
-// ----------------------------------------------------------
-
-// 设置使用这个SDK的应用程序名。userApp 必须满足 [A-Za-z0-9_\ \-\.]*
+// SetAppName 设置使用这个SDK的应用程序名。userApp 必须满足 [A-Za-z0-9_\ \-\.]*
 func SetAppName(userApp string) error {
-
-	return conf.SetAppName(userApp)
+	return conf.SetDefaultAppName(userApp)
 }
-
-// ----------------------------------------------------------
